refactor: move RC registration logic to RC product
We move some of the RC registration logic out of the global
configuration and into the RC product to remove circular import paths.
We leave it to follow-up work to improve on the modularisation of the
current solution.
